+++ web/includes/footer.php
@@ -1,11 +1,16 @@
 <div id='footer'>
-	<div class='git'>
-		<img src='images/icon/git.png'>&nbsp;<a href='https://github.com/Kubo2/diggyshelper'
-		rel='noopener' target='_blank'>GitHub <code>Kubo2/diggyshelper</code></a>
+	<div class='git' itemscope itemtype='https://schema.org/SoftwareSourceCode'>
+		<img src='images/icon/git.png'>&nbsp;<a
+			itemprop='codeRepository'
+			href='https://github.com/Kubo2/diggyshelper'
+			target='_blank'
+			rel='noopener'>GitHub <code>Kubo2/diggyshelper</code></a>
 	</div>
 	<div class='fb'>
-		<img src='images/icon/fb.png'>&nbsp;<a href='https://www.facebook.com/diggyshelper'
-		rel='noopener' target='_blank'><code>fb.com/diggyshelper</code></a>
+		<img src='images/icon/fb.png'>&nbsp;<a
+			href='https://www.facebook.com/diggyshelper'
+			target='_blank'
+			rel='noopener'><code>fb.com/diggyshelper</code></a>
 	</div>
 	<div class='copy'>
 		<span>© 2013 – <?= date('Y') ?> Diggy's Helper</span>